Split image files to image and gif
